+++ rust/legado_engine/src/frb_generated.rs
@@ -37,7 +37,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.11.1";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -2015852056;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = -1682229376;
 
 // Section: executor
 
@@ -2184,6 +2184,44 @@ fn wire__crate__api__local_book__parse_txt_chapters_impl(
         },
     )
 }
+fn wire__crate__api__dict__query_dict_rule_impl(
+    port_: flutter_rust_bridge::for_generated::MessagePort,
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_async::<flutter_rust_bridge::for_generated::SseCodec, _, _, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "query_dict_rule",
+            port: Some(port_),
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Normal,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_rule_json = <String>::sse_decode(&mut deserializer);
+            let api_word = <String>::sse_decode(&mut deserializer);
+            deserializer.end();
+            move |context| async move {
+                transform_result_sse::<_, String>(
+                    (move || async move {
+                        let output_ok =
+                            crate::api::dict::query_dict_rule(api_rule_json, api_word).await?;
+                        Ok(output_ok)
+                    })()
+                    .await,
+                )
+            }
+        },
+    )
+}
 fn wire__crate__api__record_detailed_read_session_impl(
     ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
     rust_vec_len_: i32,
@@ -3820,27 +3858,28 @@ fn pde_ffi_dispatcher_primary_impl(
         65 => {
             wire__crate__api__local_book__parse_txt_chapters_impl(port, ptr, rust_vec_len, data_len)
         }
-        67 => wire__crate__api__read_record__record_detailed_read_session_impl(
+        66 => wire__crate__api__dict__query_dict_rule_impl(port, ptr, rust_vec_len, data_len),
+        68 => wire__crate__api__read_record__record_detailed_read_session_impl(
             port,
             ptr,
             rust_vec_len,
             data_len,
         ),
-        69 => wire__crate__api__read_record__record_reading_impl(port, ptr, rust_vec_len, data_len),
-        71 => wire__crate__api__search_impl(port, ptr, rust_vec_len, data_len),
-        72 => wire__crate__api__search__search_impl(port, ptr, rust_vec_len, data_len),
-        76 => wire__crate__api__start_web_api_impl(port, ptr, rust_vec_len, data_len),
-        77 => wire__crate__api__stop_web_api_impl(port, ptr, rust_vec_len, data_len),
-        80 => wire__crate__api__validate_source_impl(port, ptr, rust_vec_len, data_len),
-        81 => wire__crate__api__validate__validate_source_impl(port, ptr, rust_vec_len, data_len),
-        83 => wire__crate__api__webdav__webdav_check_impl(port, ptr, rust_vec_len, data_len),
-        84 => wire__crate__api__webdav__webdav_delete_impl(port, ptr, rust_vec_len, data_len),
-        85 => wire__crate__api__webdav__webdav_download_impl(port, ptr, rust_vec_len, data_len),
-        86 => wire__crate__api__webdav__webdav_ensure_dir_impl(port, ptr, rust_vec_len, data_len),
-        87 => wire__crate__api__webdav__webdav_list_impl(port, ptr, rust_vec_len, data_len),
-        88 => wire__crate__api__webdav__webdav_move_impl(port, ptr, rust_vec_len, data_len),
-        89 => wire__crate__api__webdav__webdav_upload_impl(port, ptr, rust_vec_len, data_len),
-        90 => {
+        70 => wire__crate__api__read_record__record_reading_impl(port, ptr, rust_vec_len, data_len),
+        72 => wire__crate__api__search_impl(port, ptr, rust_vec_len, data_len),
+        73 => wire__crate__api__search__search_impl(port, ptr, rust_vec_len, data_len),
+        77 => wire__crate__api__start_web_api_impl(port, ptr, rust_vec_len, data_len),
+        78 => wire__crate__api__stop_web_api_impl(port, ptr, rust_vec_len, data_len),
+        81 => wire__crate__api__validate_source_impl(port, ptr, rust_vec_len, data_len),
+        82 => wire__crate__api__validate__validate_source_impl(port, ptr, rust_vec_len, data_len),
+        84 => wire__crate__api__webdav__webdav_check_impl(port, ptr, rust_vec_len, data_len),
+        85 => wire__crate__api__webdav__webdav_delete_impl(port, ptr, rust_vec_len, data_len),
+        86 => wire__crate__api__webdav__webdav_download_impl(port, ptr, rust_vec_len, data_len),
+        87 => wire__crate__api__webdav__webdav_ensure_dir_impl(port, ptr, rust_vec_len, data_len),
+        88 => wire__crate__api__webdav__webdav_list_impl(port, ptr, rust_vec_len, data_len),
+        89 => wire__crate__api__webdav__webdav_move_impl(port, ptr, rust_vec_len, data_len),
+        90 => wire__crate__api__webdav__webdav_upload_impl(port, ptr, rust_vec_len, data_len),
+        91 => {
             wire__crate__api__webdav__webdav_upload_if_match_impl(port, ptr, rust_vec_len, data_len)
         }
         _ => unreachable!(),
@@ -3899,15 +3938,15 @@ fn pde_ffi_dispatcher_sync_impl(
         61 => wire__crate__api__list_notes_impl(ptr, rust_vec_len, data_len),
         62 => wire__crate__api__parse_epub_impl(ptr, rust_vec_len, data_len),
         64 => wire__crate__api__parse_txt_chapters_impl(ptr, rust_vec_len, data_len),
-        66 => wire__crate__api__record_detailed_read_session_impl(ptr, rust_vec_len, data_len),
-        68 => wire__crate__api__record_reading_impl(ptr, rust_vec_len, data_len),
-        70 => wire__crate__api__backup__restore_backup_impl(ptr, rust_vec_len, data_len),
-        73 => wire__crate__api__seed_login_header_impl(ptr, rust_vec_len, data_len),
-        74 => wire__crate__api__network__set_network_config_impl(ptr, rust_vec_len, data_len),
-        75 => wire__crate__api__network__start_http_request_trace_impl(ptr, rust_vec_len, data_len),
-        78 => wire__crate__api__upsert_bookmark_impl(ptr, rust_vec_len, data_len),
-        79 => wire__crate__api__upsert_note_impl(ptr, rust_vec_len, data_len),
-        82 => wire__crate__api__web_api_status_impl(ptr, rust_vec_len, data_len),
+        67 => wire__crate__api__record_detailed_read_session_impl(ptr, rust_vec_len, data_len),
+        69 => wire__crate__api__record_reading_impl(ptr, rust_vec_len, data_len),
+        71 => wire__crate__api__backup__restore_backup_impl(ptr, rust_vec_len, data_len),
+        74 => wire__crate__api__seed_login_header_impl(ptr, rust_vec_len, data_len),
+        75 => wire__crate__api__network__set_network_config_impl(ptr, rust_vec_len, data_len),
+        76 => wire__crate__api__network__start_http_request_trace_impl(ptr, rust_vec_len, data_len),
+        79 => wire__crate__api__upsert_bookmark_impl(ptr, rust_vec_len, data_len),
+        80 => wire__crate__api__upsert_note_impl(ptr, rust_vec_len, data_len),
+        83 => wire__crate__api__web_api_status_impl(ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }